progress on refactoring notes details
