+++ ui/legacy.js
@@ -37,7 +37,7 @@ import '../app/assets/javascripts/student_profile/pdf/student_profile_pdf.js';
 
   // details:
 import '../app/assets/javascripts/student_profile/services_details.jsx';
-import '../app/assets/javascripts/student_profile/notes_details.jsx';
+import '../app/assets/javascripts/student_profile/notes_details.js';
 import '../app/assets/javascripts/student_profile/profile_details.jsx';
 import '../app/assets/javascripts/student_profile/attendance_details.jsx';
 import '../app/assets/javascripts/student_profile/ela_details.jsx';
